Refactor: python, cts: Move RestartTest into the _cts module.
No code changes.
