+++ python/pacemaker/_cts/tests/__init__.py
@@ -7,6 +7,7 @@ __license__ = "GNU Lesser General Public License version 2.1 or later (LGPLv2.1+
 
 from pacemaker._cts.tests.ctstest import CTSTest
 from pacemaker._cts.tests.fliptest import FlipTest
+from pacemaker._cts.tests.restarttest import RestartTest
 from pacemaker._cts.tests.remotedriver import RemoteDriver
 from pacemaker._cts.tests.simulstartlite import SimulStartLite
 from pacemaker._cts.tests.simulstoplite import SimulStopLite
